test(e2e): extract kibitz DOM-wait helpers into kibitz-helpers.ts
